Refactor "Anonymous functions"
